backend/requirements.txt
Django==5.0.9
gunicorn==22.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
djangorestframework==3.15.2
django-modeltranslation==0.18.11
cryptography===43.0.1